chore(protobuf): rename `anchor` dir to `protobuf`
