+++ anchor/.env.sample
@@ -1,0 +1,2 @@
+PRIVATE_KEY__BCM_IIUM=some-private-key
+PRIVATE_KEY__BCM_NEM=some-private-key
