+++ anchor/.gitignore
@@ -1,0 +1,3 @@
+anchors.json
+node_modules
+.now
