+++ island/nem/package.json
@@ -8,7 +8,7 @@
   "license": "Apache-2.0",
   "scripts": {
     "build": "tsc",
-    "dev": "ts-node index.ts",
+    "dev": "ts-node cli.ts",
     "proto": "yarn run proto-generate && yarn run proto-generate-types",
     "proto-generate": "grpc_tools_node_protoc --js_out=import_style=commonjs,binary:./anchor --grpc_out=./anchor --plugin=protoc-gen-grpc=`which grpc_tools_node_protoc_plugin` -I ../../anchor anchor.proto",
     "proto-generate-types": "protoc --plugin=protoc-gen-ts=./node_modules/.bin/protoc-gen-ts --ts_out=./anchor -I ../../anchor anchor.proto"
